langgraph-skills/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "langgraph-skills"
version = "0.1.0"
description = "LangGraph-based dynamic skills system"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.1.0",
    "langchain-core>=0.2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
